@@ -28,17 +28,15 @@
     "babel-preset-stage-1": "^6.16.0",
     "babel-register": "^6.18.0",
     "classnames": "^2.2.5",
-    "clean-webpack-plugin": "^0.1.14",
-    "copy-to-clipboard": "^3.0.5",
-    "css-loader": "^0.26.1",
+    "clean-webpack-plugin": "^1.0.1",
+    "copy-to-clipboard": "^3.0.8",
+    "css-loader": "^2.1.0",
     "dotenv": "^2.0.0",
     "entities": "^1.1.1",
     "eslint": "^4.19.1",
-    "extract-text-webpack-plugin": "^3.0.2",
     "favicons-webpack-plugin": "^0.0.8",
-    "file-loader": "^0.9.0",
+    "file-loader": "^3.0.1",
     "file-saver": "^1.3.8",
-    "font-awesome": "^4.7.0",
     "foundation-sites": "^6.3.0",
     "fraction.js": "^4.0.0",
     "fuse.js": "^2.6.2",
@@ -51,12 +49,11 @@
     "leaflet-geosearch": "^2.6.0",
     "leaflet-measure-path": "^1.3.1",
     "leaflet-zoombox": "^0.5.1",
-    "line-height": "^0.3.1",
     "lodash": "^4.17.2",
     "material-icons": "^0.1.0",
+    "mini-css-extract-plugin": "^0.5.0",
     "moment": "^2.20.1",
-    "morgan": "^1.8.1",
-    "node-sass": "^4.9.0",
+    "node-sass": "^4.11.0",
     "normalize.css": "^5.0.0",
     "numeral": "^2.0.6",
     "offline-plugin": "^4.5.3",
@@ -64,11 +61,10 @@
     "proj4": "^2.4.3",
     "proj4leaflet": "^1.0.1",
     "prop-types": "^15.6.0",
-    "react": "^16.5.2",
+    "react": "^16.7.0",
     "react-datepicker": "^1.4.1",
-    "react-dom": "^16.5.2",
+    "react-dom": "^16.7.0",
     "react-dual-listbox": "^1.3.2",
-    "react-element-to-jsx-string": "^13.2.0",
     "react-foundation": "^0.9.5",
     "react-hot-loader": "^3.1.3",
     "react-i18next": "^2.2.0",
@@ -79,21 +75,21 @@
     "react-redux": "^5.0.7",
     "react-redux-toastr": "^7.4.0",
     "react-resize-detector": "^2.3.0",
-    "react-router": "^3.0.0",
+    "react-router": "^4.3.1",
     "react-router-redux": "^4.0.8",
     "react-scroll-to-component": "^1.0.2",
     "react-select": "^1.3.0",
     "react-transition-group": "^2.3.1",
-    "redux": "^3.6.0",
-    "redux-actions": "^1.2.1",
+    "redux": "^4.0.1",
+    "redux-actions": "^2.6.4",
     "redux-form": "^7.4.2",
-    "redux-oidc": "^3.0.0-beta.7",
-    "redux-saga": "^0.16.0",
+    "redux-oidc": "^3.1.0",
+    "redux-saga": "^1.0.0",
     "sass-loader": "^7.0.1",
-    "style-loader": "^0.13.1",
+    "uglifyjs-webpack-plugin": "^2.1.1",
     "url-loader": "^0.5.7",
-    "webpack": "^3.0.0",
-    "webpack-merge": "^1.1.2",
+    "webpack": "^4.4.0",
+    "webpack-merge": "^4.2.1",
     "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
@@ -107,7 +103,7 @@
     "debug": "^2.6.3",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
-    "eslint-plugin-react": "^7.7.0",
+    "eslint-plugin-react": "^7.12.4",
     "flow-bin": "^0.71",
     "flow-typed": "^2.4.0",
     "husky": "^0.12.0",
@@ -115,13 +111,13 @@
     "istanbul": "v1.1.0-alpha.1",
     "mocha": "^3.2.0",
     "nodemon": "^1.17.4",
-    "npm-run-all": "^3.1.2",
-    "postcss-loader": "^1.2.2",
+    "npm-run-all": "^4.1.5",
     "sinon": "^4.0.0",
     "sinon-chai": "^3.0.0",
-    "webpack-cli": "^1.3.2",
-    "webpack-dashboard": "^0.3.0",
-    "webpack-dev-server": "^2.11.1",
-    "webpack-hot-middleware": "^2.22.1"
+    "style-loader": "^0.23.1",
+    "webpack-cli": "^3.2.1",
+    "webpack-dashboard": "^2.0.0",
+    "webpack-dev-server": "^3.1.14",
+    "webpack-hot-middleware": "^2.24.3"
   }
 }
